Moved Gen functions from EclTestData into Generators
