+++ test-common/uk/gov/hmrc/economiccrimelevyreturns/generators/Generators.scala
@@ -19,6 +19,9 @@ package uk.gov.hmrc.economiccrimelevyreturns.generators
 import org.scalacheck.Arbitrary._
 import org.scalacheck.Gen._
 import org.scalacheck.{Gen, Shrink}
+import org.scalacheck.Gen.{choose, listOfN}
+import uk.gov.hmrc.economiccrimelevyreturns.Regex
+import wolfendale.scalacheck.regexp.RegexpGen
 
 import java.time.{Instant, LocalDate, ZoneOffset}
 
@@ -85,6 +88,12 @@ trait Generators {
       chars  <- listOfN(length, arbitrary[Char])
     } yield chars.mkString
 
+  def alphaNumStringsWithMaxLength(maxLength: Int): Gen[String] =
+    for {
+      length <- choose(1, maxLength)
+      chars  <- listOfN(length, Gen.alphaNumChar)
+    } yield chars.mkString
+
   def stringsLongerThan(minLength: Int): Gen[String] = for {
     maxLength <- (minLength * 2).max(100)
     length    <- Gen.chooseNum(minLength + 1, maxLength)
@@ -111,4 +120,17 @@ trait Generators {
       Instant.ofEpochMilli(millis).atOffset(ZoneOffset.UTC).toLocalDate
     }
   }
+
+  def emailAddress(maxLength: Int): Gen[String] = {
+    val emailPartsLength = maxLength / 5
+
+    for {
+      firstPart  <- alphaNumStringsWithMaxLength(emailPartsLength)
+      secondPart <- alphaNumStringsWithMaxLength(emailPartsLength)
+      thirdPart  <- alphaNumStringsWithMaxLength(emailPartsLength)
+    } yield s"$firstPart@$secondPart.$thirdPart".toLowerCase
+  }
+
+  def telephoneNumber(maxLength: Int): Gen[String] =
+    RegexpGen.from(s"${Regex.telephoneNumber}").retryUntil(s => s.length <= maxLength && s.trim.nonEmpty)
 }
